refactor, creating supervisor submodule
