+++ meta-bkk-setup/recipes-setup/application-manager/files/src/am_supervisor.h
@@ -1,0 +1,10 @@
+#ifndef AM_SUPERVISOR_H
+#define AM_SUPERVISOR_H
+
+#include "am_types.h"
+
+int start_supervisor(app_info_t * app_infos, int num_apps);
+
+
+
+#endif // AM_SUPERVISOR_H
